Add Eslint and Stylelint
