+++ .gitignore
@@ -33,3 +33,6 @@ yarn-error.log*
 
 # vercel
 .vercel
+
+#vscode
+.vscode
