+++ package.json
@@ -5,7 +5,12 @@
   "scripts": {
     "dev": "next dev",
     "build": "next build",
-    "start": "next start"
+    "start": "next start",
+    "lint": "npm-run-all lint:js lint:css",
+    "lint:css": "stylelint **/*.scss --syntax scss || true",
+    "lint:css:fix": "stylelint src/**/*.scss --fix",
+    "lint:js": "eslint src/",
+    "lint:js:fix": "eslint src/ --fix"
   },
   "dependencies": {
     "@popperjs/core": "^2.9.1",
@@ -17,5 +22,34 @@
     "react": "17.0.1",
     "react-dom": "17.0.1",
     "react-to-print": "^2.12.3"
+  },
+  "devDependencies": {
+    "@babel/core": "^7.13.10",
+    "@babel/plugin-transform-runtime": "^7.13.10",
+    "@babel/preset-env": "^7.13.12",
+    "@babel/preset-react": "^7.12.13",
+    "babel-jest": "^26.6.3",
+    "babel-plugin-dynamic-import-node": "^2.3.3",
+    "eslint": "^7.22.0",
+    "eslint-config-airbnb": "^18.2.1",
+    "eslint-plugin-import": "^2.22.1",
+    "eslint-plugin-jsx-a11y": "^6.4.1",
+    "eslint-plugin-react": "^7.23.1",
+    "eslint-plugin-react-hooks": "^4.2.0",
+    "husky": "^5.2.0",
+    "npm-run-all": "^4.1.5",
+    "stylelint": "^13.12.0",
+    "stylelint-config-sass-guidelines": "^8.0.0",
+    "stylelint-config-standard": "^21.0.0",
+    "stylelint-order": "^4.1.0"
+  },
+  "eslintConfig": {
+    "extends": "react-app"
+  },
+  "husky": {
+    "hooks": {
+      "pre-commit": "npm run lint && npm run test",
+      "pre-push": "npm run build"
+    }
   }
 }
