+++ src/pages/_app.js
@@ -1,5 +1,5 @@
-import "bootstrap/dist/css/bootstrap.css";
-import "../styles/globals.scss";
+import 'bootstrap/dist/css/bootstrap.css';
+import '../styles/globals.scss';
 
 function MyApp({ Component, pageProps }) {
   return <Component {...pageProps} />;
