+++ src/styles/globals.scss
@@ -1,0 +1,6 @@
+@import './fonts.scss';
+@import './main.scss';
+
+body {
+  font-family: 'Arvo-Regular';
+}
