+++ src/styles/main.scss
@@ -9,25 +9,25 @@
 
   .profile {
     .profile-item {
+      border-bottom: 4px solid #d3d3d3;
       margin: 40px 40px 0;
       padding-bottom: 40px;
-      border-bottom: 4px solid #d3d3d3;
     }
   }
 
   h3 {
-    position: relative;
     margin-bottom: 32px;
+    position: relative;
 
     &::after {
-      background-color: #000;
+      background-color: #000000;
       border-radius: 8px;
-      content: "";
+      content: '';
       height: 6px;
-      position: absolute;
-      width: 40px;
       left: 0;
+      position: absolute;
       top: 50px;
+      width: 40px;
     }
   }
 }
