--- styles/globals.scss
@@ -1,6 +1,0 @@
-@import "./fonts.scss";
-@import "./main.scss";
-
-body {
-  font-family: "Arvo-Regular";
-}
